+++ common/src/main/java/com/majruszsdifficulty/entity/CerberusRenderer.java
@@ -16,12 +16,14 @@ import net.minecraft.resources.ResourceLocation;
 @OnlyIn( Dist.CLIENT )
 public class CerberusRenderer extends MobRenderer< Cerberus, CerberusModel< Cerberus > > {
 	public static final ModelLayerLocation LAYER = MajruszsDifficulty.HELPER.getLayerLocation( "cerberus" );
+	public static final ModelLayerLocation ARMOR_LAYER = MajruszsDifficulty.HELPER.getLayerLocation( "cerberus", "armor" );
 	public static final ResourceLocation TEXTURE = MajruszsDifficulty.HELPER.getLocation( "textures/entity/cerberus.png" );
 
 	public CerberusRenderer( EntityRendererProvider.Context context ) {
 		super( context, new CerberusModel<>( context.bakeLayer( LAYER ) ), 0.75f );
 
 		this.addLayer( new CerberusEyesLayer( this ) );
+		this.addLayer( new UndeadArmyArmorLayer<>( this, new CerberusModel<>( context.bakeLayer( ARMOR_LAYER ) ), "textures/entity/cerberus_undead_army_armor.png" ) );
 	}
 
 	@Override